# schemas/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rideshare Simulation Control Panel API",
    "description": "REST API for controlling simulation and streaming real-time updates",
    "version": "1.0.0"
  },
  "paths": {
...
    "/agents/drivers": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create Drivers",
        "description": "Queue driver agents for continuous spawning.\n\nDrivers are spawned at a continuous rate (default: 2/sec) to prevent\nsynchronized GPS ping bursts. Use GET /agents/spawn-status to monitor progress.\n\nQuery Parameters:\n    mode: immediate (default) - drivers go online immediately\n          scheduled - drivers follow their DNA shift_preference schedule",
        "operationId": "create_drivers_agents_drivers_post",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SpawnMode",
              "description": "immediate: go online immediately; scheduled: follow DNA shift_preference",
              "default": "immediate"
            },
            "description": "immediate: go online immediately; scheduled: follow DNA shift_preference"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriversCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/riders": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create Riders",
        "description": "Queue rider agents for continuous spawning.\n\nRiders are spawned at a continuous rate (default: 40/sec) to prevent\nsynchronized GPS ping bursts. Use GET /agents/spawn-status to monitor progress.\n\nQuery Parameters:\n    mode: immediate - riders request a trip immediately after spawning\n          scheduled (default) - riders follow their DNA avg_rides_per_week schedule",
        "operationId": "create_riders_agents_riders_post",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SpawnMode",
              "description": "immediate: request trip immediately; scheduled: follow DNA avg_rides_per_week",
              "default": "scheduled"
            },
            "description": "immediate: request trip immediately; scheduled: follow DNA avg_rides_per_week"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiderCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RidersCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/spawn-status": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Spawn Status",
        "description": "Get current spawn queue status.\n\nReturns the number of drivers and riders waiting to be spawned.",
        "operationId": "get_spawn_status_agents_spawn_status_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpawnQueueStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/drivers/{driver_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Driver State",
        "description": "Get full state and DNA for a driver.\n\nReturns the driver's current state (status, location, rating),\nactive trip info if any, and complete DNA parameters.",
        "operationId": "get_driver_state_agents_drivers__driver_id__get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/riders/{rider_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Rider State",
        "description": "Get full state and DNA for a rider.\n\nReturns the rider's current state (status, location, rating),\nactive trip info if any, and complete DNA parameters.",
        "operationId": "get_rider_state_agents_riders__rider_id__get",
        "parameters": [
          {
            "name": "rider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rider Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiderStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/drivers/{driver_id}/status": {
      "put": {
        "tags": [
          "agents"
        ],
        "summary": "Toggle Driver Status",
        "description": "Toggle driver online/offline status.\n\nCan only toggle when driver has no active trip.",
        "operationId": "toggle_driver_status_agents_drivers__driver_id__status_put",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverStatusToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverStatusToggleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/riders/{rider_id}/request-trip": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Request Rider Trip",
        "description": "Request a trip for a rider with specified destination.\n\nThe rider must be in 'offline' status to request a trip.",
        "operationId": "request_rider_trip_agents_riders__rider_id__request_trip_post",
        "parameters": [
          {
            "name": "rider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rider Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiderTripRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiderTripRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriverCreateRequest": {
        "properties": {
          "count": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "DriverCreateRequest"
      },
...
      "DriverStateResponse": {
        "properties": {
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "location": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "current_rating": {
            "type": "number",
            "title": "Current Rating"
          },
          "rating_count": {
            "type": "integer",
            "title": "Rating Count"
          },
          "active_trip": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActiveTripInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_offer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingOfferInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NextActionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "zone_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Id"
          },
          "dna": {
            "$ref": "#/components/schemas/DriverDNAResponse"
          },
          "statistics": {
            "$ref": "#/components/schemas/DriverStatisticsResponse"
          },
          "is_ephemeral": {
            "type": "boolean",
            "title": "Is Ephemeral"
          },
          "is_puppet": {
            "type": "boolean",
            "title": "Is Puppet"
          }
        },
        "type": "object",
        "required": [
          "driver_id",
          "status",
          "location",
          "current_rating",
          "rating_count",
          "active_trip",
          "zone_id",
          "dna",
          "statistics",
          "is_ephemeral",
          "is_puppet"
        ],
        "title": "DriverStateResponse",
        "description": "Full driver state for inspection."
      },
...
      "DriverStatusToggleRequest": {
        "properties": {
          "go_online": {
            "type": "boolean",
            "title": "Go Online",
            "description": "True to go online, False to go offline"
          }
        },
        "type": "object",
        "required": [
          "go_online"
        ],
        "title": "DriverStatusToggleRequest",
        "description": "Request to toggle driver online/offline status."
      },
...
      "DriversCreateResponse": {
        "properties": {
          "queued": {
            "type": "integer",
            "title": "Queued"
          },
          "spawn_rate": {
            "type": "number",
            "title": "Spawn Rate"
          },
          "estimated_completion_seconds": {
            "type": "number",
            "title": "Estimated Completion Seconds"
          }
        },
        "type": "object",
        "required": [
          "queued",
          "spawn_rate",
          "estimated_completion_seconds"
        ],
        "title": "DriversCreateResponse",
        "description": "Response for queued driver creation with continuous spawning."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiderCreateRequest": {
        "properties": {
          "count": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "RiderCreateRequest"
      },
...
      "RidersCreateResponse": {
        "properties": {
          "queued": {
            "type": "integer",
            "title": "Queued"
          },
          "spawn_rate": {
            "type": "number",
            "title": "Spawn Rate"
          },
          "estimated_completion_seconds": {
            "type": "number",
            "title": "Estimated Completion Seconds"
          }
        },
        "type": "object",
        "required": [
          "queued",
          "spawn_rate",
          "estimated_completion_seconds"
        ],
        "title": "RidersCreateResponse",
        "description": "Response for queued rider creation with continuous spawning."
      },
...
      "SpawnMode": {
        "type": "string",
        "enum": [
          "immediate",
          "scheduled"
        ],
        "title": "SpawnMode",
        "description": "Spawn mode for autonomous agents."
      },
      "SpawnQueueStatusResponse": {
        "properties": {
          "drivers_queued": {
            "type": "integer",
            "title": "Drivers Queued"
          },
          "riders_queued": {
            "type": "integer",
            "title": "Riders Queued"
          }
        },
        "type": "object",
        "required": [
          "drivers_queued",
          "riders_queued"
        ],
        "title": "SpawnQueueStatusResponse",
        "description": "Current status of agent spawn queues."
      },
...
    }
  }
}